Add new files and update existing ones for testnet and logstore
This commit adds new files for setting up the blockchain testnet and updates existing files related to logstore. This includes addressing a change in the authentication function, switching time calculations from Unix to UnixMilli, and modifying the LogStoreClient to handle fetching messages using requests. Additionally, this commit introduces the 'gotest.tools' dependency and updates the docker-compose files accordingly. Other minor changes include updates to genesis.json and private_key files for the testnet, as well as fixes in the Go module requirements.