Update Docker image and extend README instructions
The Docker image source has been updated to pull from the 'usherlabs/logstore-node' repository, in specific from the 'develop' tag. The README file has been expanded with a comprehensive guide detailing how to run the application, which includes steps from installing pre-requisites to deploying the demo schema and verifying output. This is particularly beneficial for increasing user-friendliness and accessibility of the code.