Update gitignore and docker-compose file paths
The changes in the commit focus on updating a couple of file paths both in the .gitignore and docker-compose.yaml files. The modifications in the .gitignore file involve changing the ignored directory for data storage related to examples. The docker-compose.yaml file has had paths modified to match their correct locations, specifically for configuration and private key files.