Update docker images to dev version and improve README instructions
The docker-compose files have been updated to use the development version of the logstore nodes. The README file has been also updated with added detail on testing a single node, including what to do in case of network instability and reminders for setting up a wallet if not configured yet.